+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The main focus is on user-observable behavior of the engine.
 ## Version 1.0.0 RC4
 * Added stack trace limit option (--js.stack-trace-limit).
 * Enable SharedArrayBuffers by default.
+* Provide $EXEC for Nashorn compatibility in scripting mode.
 
 ## Version 1.0.0 RC3
 * Added support for BigInt arithmetic expressions.